arm64: dt: tqma8qxp-mba8xx: disable voltage change for sd-card
Because there is no posibility on mainboard to switch card supply voltage.
there is no way to reset the card after after beeing in UHS mode (ultra high speed mode).

To be able to do testing with UHS, add a special device tree
that enables switching the IO voltage for the SD card.

Signed-off-by: Martin Schmiedel <martin.schmiedel@tq-group.com>