arm64: dt: imx8qxp-mba8xx: remove local-mac-address prop from fec
mac-address has the highest priority per binding doc and should contain
the last used address. This one should be transferred to the kernel in
to make sure the MAC changes not only due to booting. Bootloaders are
free to add the local-mac-address using the real hardware supplied address
in case it differs from the last used address.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>