arm64: dt: imx8mq-mba8mx: add GPIO for SD card supply
With mainboard rev.030x the SD RESET signal ca be used to switch of
the SD card supply. Athough this is more important for other SOM using
this board, the GPIO should be under control of the SD card driver to
prevent irregular behavour when using the signal as normal GPIO.

Since the signal has no meaning and is not connecte for HW rev.0200
we do not need to take special care for the older revision.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>